refactor: Extract feedback/survey cards into separate components
